--- web/src/app/course-table/admin-schedule/admin-schedule.component.html
@@ -1,58 +1,0 @@
-<nav class="navbar navbar-light bg-light">
-  <form class="form-inline">
-    <div class="mr-sm-2">
-      <app-school-select [(ngModel)]="selectedSchool" (ngModelChange)="onSchoolChange($event)" name="selectedSchool">
-      </app-school-select>
-    </div>
-    <div class="mr-sm-2">
-      <nz-select nzPlaceHolder="选择周次"
-                 [(ngModel)]="selectedWeek"
-                 name="selectedWeek">
-        <nz-option  *ngFor="let week of allWeeks" [nzValue]="week" [nzLabel]="'第' + week + '周'"></nz-option>
-      </nz-select>
-    </div>
-    <div class="mr-sm-2">
-      <nz-select nzPlaceHolder="选择学生" nzAllowClear
-                 [(ngModel)]="selectedStudent"
-                 name="selectedStudent">
-        <nz-option  *ngFor="let student of allStudents" [nzValue]="student.id" [nzLabel]="student.name"></nz-option>
-      </nz-select>
-    </div>
-    <button class="btn btn-outline-success my-2 my-sm-0" (click)="getCourseMessage()">搜索</button>
-  </form>
-</nav>
-
-<div class="schedule-container">
-  <p>当前学期为：{{ termMessage.name }}</p>
-  <table class="table table-bordered">
-    <thead>
-    <tr>
-      <th>时间</th>
-      <th>周一</th>
-      <th>周二</th>
-      <th>周三</th>
-      <th>周四</th>
-      <th>周五</th>
-      <th>周六</th>
-      <th>周日</th>
-    </tr>
-    </thead>
-    <tbody>
-    <!-- 遍历时间段 -->
-    <tr *ngFor="let period of periods">
-      <td>{{ period.time }}</td>
-
-      <!-- 遍历每天的课程信息 -->
-      <ng-container *ngFor="let day of days">
-        <td *ngIf="shouldShowCell(day, period.time)"
-            [attr.rowspan]="getClassInfo(day, period.time)?.length || 1">
-          <ng-container *ngIf="getClassInfo(day, period.time) as classInfo; else noCourse">
-            <p>{{ classInfo.students }}</p>
-          </ng-container>
-          <ng-template #noCourse></ng-template>
-        </td>
-      </ng-container>
-    </tr>
-    </tbody>
-  </table>
-</div>
