--- web/src/app/course-table/admin-schedule/admin-schedule.component.spec.ts
@@ -1,36 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { AdminScheduleComponent } from './admin-schedule.component';
-import {HttpClientTestingModule} from "@angular/common/http/testing";
-import {SchoolSelectModule} from "../../core/school-select/school-select.module";
-import {NzSelectModule} from "ng-zorro-antd/select";
-import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
-
-describe('AdminScheduleComponent', () => {
-  let component: AdminScheduleComponent;
-  let fixture: ComponentFixture<AdminScheduleComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ AdminScheduleComponent ],
-      imports: [
-        HttpClientTestingModule,
-        SchoolSelectModule,
-        NzSelectModule,
-        FormsModule,
-        ReactiveFormsModule,
-        BrowserAnimationsModule
-      ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(AdminScheduleComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
