+++ web/src/app/course-table/course-table.component.spec.ts
@@ -1,14 +1,11 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { CourseTableComponent } from './course-table.component';
-import {AdminScheduleComponent} from "./admin-schedule/admin-schedule.component";
 import {HttpClientTestingModule} from "@angular/common/http/testing";
 import {SchoolSelectModule} from "../core/school-select/school-select.module";
-import {SchoolSelectComponent} from "../core/school-select/school-select.component";
 import {NzSelectModule} from "ng-zorro-antd/select";
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
-import {StudentScheduleComponent} from "./student-schedule/student-schedule.component";
 
 describe('CourseTableComponent', () => {
   let component: CourseTableComponent;
@@ -16,7 +13,7 @@ describe('CourseTableComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ CourseTableComponent, AdminScheduleComponent, StudentScheduleComponent ],
+      declarations: [ CourseTableComponent ],
       imports: [
         HttpClientTestingModule,
         SchoolSelectModule,
