+++ web/src/app/course-table/course-table.module.ts
@@ -3,15 +3,13 @@ import { CommonModule } from '@angular/common';
 
 import { CourseTableRoutingModule } from './course-table-routing.module';
 import {CourseTableComponent} from "./course-table.component";
-import { AdminScheduleComponent } from './admin-schedule/admin-schedule.component';
 import {SchoolSelectModule} from "../core/school-select/school-select.module";
 import {NzSelectModule} from "ng-zorro-antd/select";
 import {FormsModule} from "@angular/forms";
-import { StudentScheduleComponent } from './student-schedule/student-schedule.component';
 
 
 @NgModule({
-  declarations: [CourseTableComponent, AdminScheduleComponent, StudentScheduleComponent],
+  declarations: [CourseTableComponent],
   imports: [
     CommonModule,
     CourseTableRoutingModule,
