--- web/src/app/course-table/student-schedule/student-schedule.component.css
@@ -1,29 +1,0 @@
-.btn {
-  margin-right: 0.5em;
-  margin-top: 0.7em;
-}
-
-.btn > svg.svg-inline--fa {
-  margin-right: 0.3em;
-}.schedule-container {
-   margin: 20px;
-   font-family: Arial, sans-serif;
- }
-
-.table {
-  width: 100%;
-  text-align: center;
-}
-
-.table th, .table td {
-  padding: 10px;
-  border: 1px solid #ddd;
-}
-
-.table th {
-  background-color: #f4f4f4;
-}
-
-.table td {
-  vertical-align: middle;
-}
