--- web/src/app/course-table/student-schedule/student-schedule.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { StudentScheduleComponent } from './student-schedule.component';
-import {HttpClientTestingModule} from "@angular/common/http/testing";
-
-describe('StudentScheduleComponent', () => {
-  let component: StudentScheduleComponent;
-  let fixture: ComponentFixture<StudentScheduleComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ StudentScheduleComponent ],
-      imports: [ HttpClientTestingModule ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(StudentScheduleComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
